Fix `stable_sort_primitive` to prefer `sort` not only primitive types
